moving headers in SiStripDetId and TrackerCommon to avoid circular dependencies
