+++ DQM/SiStripMonitorSummary/src/SiStripBackPlaneCorrectionDQM.cc
@@ -1,6 +1,6 @@
 #include "DQM/SiStripMonitorSummary/interface/SiStripBackPlaneCorrectionDQM.h"
 #include "DQMServices/Core/interface/DQMStore.h"
-#include "DataFormats/SiStripDetId/interface/SiStripSubStructure.h"
+#include "DataFormats/TrackerCommon/interface/SiStripSubStructure.h"
 #include "Geometry/Records/interface/TrackerTopologyRcd.h"
 #include "TCanvas.h"
